fix(issue-9 & issue-10): move project package and include file header
