--- api/src/main/kotlin/io/github/luizotavio/exception/InternalPlaceholderException.kt
@@ -1,7 +1,0 @@
-package io.github.luizotavio.exception
-
-class InternalPlaceholderException : Exception {
-    constructor(throwable: Throwable) : super(throwable)
-
-    constructor(message: String) : super(message)
-}
