+++ gradle.properties
@@ -1,4 +1,4 @@
 kotlin.code.style=official
 
 placeholder_version=1.1.0
-placeholder_groupId=io.github.luizotavio
+placeholder_groupId=me.luizotavio.minecraft
