+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'wisply-placeholder'
 
-include 'api'
-include 'common'
-include 'test'
-
+include(
+    'api',
+    'common',
+    'test'
+)
