+++ test/build.gradle
@@ -2,8 +2,8 @@ plugins {
     id 'java'
 }
 
-group 'io.github.luizotavio'
-version '1.0.0-SNAPSHOT'
+group placeholder_groupId
+version placeholder_version
 
 repositories {
     mavenCentral()
@@ -11,7 +11,7 @@ repositories {
 
 dependencies {
     testImplementation(project(":api"))
-    
+
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
 }
